Refactor common code to abstract CentralProcessor class
