+++ src/main/java/oshi/hardware/platform/linux/LinuxCentralProcessor.java
@@ -20,28 +20,20 @@ package oshi.hardware.platform.linux;
 import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
-import java.lang.management.ManagementFactory;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-
-import javax.json.Json;
-import javax.json.JsonArrayBuilder;
-import javax.json.JsonBuilderFactory;
-import javax.json.JsonObject;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.sun.jna.Native;
 
-import oshi.hardware.CentralProcessor;
+import oshi.hardware.common.AbstractCentralProcessor;
 import oshi.jna.platform.linux.Libc;
 import oshi.jna.platform.linux.Libc.Sysinfo;
-import oshi.json.NullAwareJsonObjectBuilder;
 import oshi.util.ExecutingCommand;
 import oshi.util.FileUtil;
 import oshi.util.ParseUtil;
@@ -53,102 +45,78 @@ import oshi.util.ParseUtil;
  * @author alessio.fachechi[at]gmail[dot]com
  * @author widdis[at]gmail[dot]com
  */
-@SuppressWarnings("restriction")
-public class LinuxCentralProcessor implements CentralProcessor {
+public class LinuxCentralProcessor extends AbstractCentralProcessor {
 
     private static final Logger LOG = LoggerFactory.getLogger(LinuxCentralProcessor.class);
-
-    // Determine whether MXBean supports Oracle JVM methods
-    private static final java.lang.management.OperatingSystemMXBean OS_MXBEAN = ManagementFactory
-            .getOperatingSystemMXBean();
-
-    private static boolean sunMXBean;
-
-    static {
-        try {
-            Class.forName("com.sun.management.OperatingSystemMXBean");
-            // Initialize CPU usage
-            ((com.sun.management.OperatingSystemMXBean) OS_MXBEAN).getSystemCpuLoad();
-            sunMXBean = true;
-            LOG.debug("Oracle MXBean detected.");
-        } catch (ClassNotFoundException e) {
-            sunMXBean = false;
-            LOG.debug("Oracle MXBean not detected.");
-            LOG.trace("", e);
-        }
-    }
-
-    // Logical and Physical Processor Counts
-    private int logicalProcessorCount = 0;
-
-    private int physicalProcessorCount = 0;
-
-    // Maintain previous ticks to be used for calculating usage between them.
-    // System ticks
-    private long tickTime;
-
-    private long[] prevTicks;
-
-    private long[] curTicks;
-
-    // Per-processor ticks [cpu][type]
-    private long procTickTime;
-
-    private long[][] prevProcTicks;
-
-    private long[][] curProcTicks;
-
-    // Processor info
-    private String cpuVendor;
-
-    private String cpuName;
-
-    private String cpuSerialNumber = null;
-
-    private String cpuIdentifier;
-
-    private String cpuStepping;
-
-    private String cpuModel;
-
-    private String cpuFamily;
-
-    private Long cpuVendorFreq;
-
-    private Boolean cpu64;
-
-    private JsonBuilderFactory jsonFactory = Json.createBuilderFactory(null);
 
     /**
      * Create a Processor
      */
     public LinuxCentralProcessor() {
-        // Processor counts
-        calculateProcessorCounts();
-
-        // System ticks
-        this.prevTicks = new long[4];
-        this.curTicks = new long[4];
-        updateSystemTicks();
-
-        // Per-processor ticks
-        this.prevProcTicks = new long[logicalProcessorCount][4];
-        this.curProcTicks = new long[logicalProcessorCount][4];
-        updateProcessorTicks();
+        // Initialize class variables
+        initVars();
+        // Initialize tick arrays
+        initTicks();
 
         LOG.debug("Initialized Processor");
+    }
+
+    private void initVars() {
+        List<String> cpuInfo = null;
+        try {
+            cpuInfo = FileUtil.readFile("/proc/cpuinfo");
+        } catch (IOException e) {
+            LOG.error("Problem with /proc/cpuinfo: {}", e.getMessage());
+            return;
+        }
+        for (String line : cpuInfo) {
+            String[] splitLine = line.split("\\s+:\\s");
+            if (splitLine.length < 2) {
+                break;
+            }
+            switch (splitLine[0]) {
+            case "vendor_id":
+                this.setVendor(splitLine[1]);
+                break;
+            case "model name":
+                this.setName(splitLine[1]);
+                break;
+            case "flags":
+                String[] flags = splitLine[1].toUpperCase().split(" ");
+                boolean found = false;
+                for (String flag : flags) {
+                    if (flag.equals("LM")) {
+                        found = true;
+                        break;
+                    }
+                }
+                this.setCpu64(found);
+                break;
+            case "stepping":
+                this.setStepping(splitLine[1]);
+                break;
+            case "model":
+                this.setModel(splitLine[1]);
+                break;
+            case "cpu family":
+                this.setFamily(splitLine[1]);
+                break;
+            default:
+                // Do nothing
+            }
+        }
     }
 
     /**
      * Updates logical and physical processor counts from /proc/cpuinfo
      */
-    private void calculateProcessorCounts() {
+    protected void calculateProcessorCounts() {
         try {
             List<String> procCpu = FileUtil.readFile("/proc/cpuinfo");
             // Get number of logical processors
             for (String cpu : procCpu) {
                 if (cpu.startsWith("processor")) {
-                    logicalProcessorCount++;
+                    this.logicalProcessorCount++;
                 }
             }
             // Get number of physical processors
@@ -165,7 +133,7 @@ public class LinuxCentralProcessor implements CentralProcessor {
                     // if siblings = 1, no hyperthreading
                     siblings = ParseUtil.parseString(cpu, 1);
                     if (siblings == 1) {
-                        physicalProcessorCount = logicalProcessorCount;
+                        this.physicalProcessorCount = this.logicalProcessorCount;
                         break;
                     }
                 }
@@ -173,7 +141,7 @@ public class LinuxCentralProcessor implements CentralProcessor {
                     // if siblings > 1, ratio with cores
                     cpucores = ParseUtil.parseString(cpu, 1);
                     if (siblings > 1) {
-                        physicalProcessorCount = logicalProcessorCount * cpucores / siblings;
+                        this.physicalProcessorCount = this.logicalProcessorCount * cpucores / siblings;
                         break;
                     }
                 }
@@ -190,200 +158,21 @@ public class LinuxCentralProcessor implements CentralProcessor {
                     uniqueID[1] = -1;
                 }
             }
-            if (physicalProcessorCount == 0) {
-                physicalProcessorCount = ids.size();
+            if (this.physicalProcessorCount == 0) {
+                this.physicalProcessorCount = ids.size();
             }
         } catch (IOException e) {
             LOG.error("Problem with /proc/cpuinfo: {}", e.getMessage());
         }
         // Force at least one processor
-        if (logicalProcessorCount < 1) {
+        if (this.logicalProcessorCount < 1) {
             LOG.error("Couldn't find logical processor count. Assuming 1.");
-            logicalProcessorCount = 1;
+            this.logicalProcessorCount = 1;
         }
-        if (physicalProcessorCount < 1) {
+        if (this.physicalProcessorCount < 1) {
             LOG.error("Couldn't find physical processor count. Assuming 1.");
-            physicalProcessorCount = 1;
+            this.physicalProcessorCount = 1;
         }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getVendor() {
-        return this.cpuVendor;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setVendor(String vendor) {
-        this.cpuVendor = vendor;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getName() {
-        return this.cpuName;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setName(String name) {
-        this.cpuName = name;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public long getVendorFreq() {
-        if (this.cpuVendorFreq == null) {
-            Pattern pattern = Pattern.compile("@ (.*)$");
-            Matcher matcher = pattern.matcher(getName());
-
-            if (matcher.find()) {
-                String unit = matcher.group(1);
-                this.cpuVendorFreq = Long.valueOf(ParseUtil.parseHertz(unit));
-            } else {
-                this.cpuVendorFreq = Long.valueOf(-1L);
-            }
-        }
-
-        return this.cpuVendorFreq.longValue();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setVendorFreq(long freq) {
-        this.cpuVendorFreq = Long.valueOf(freq);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getIdentifier() {
-        if (this.cpuIdentifier == null) {
-            StringBuilder sb = new StringBuilder();
-            if (getVendor() != null && getVendor().contentEquals("GenuineIntel")) {
-                sb.append(isCpu64bit() ? "Intel64" : "x86");
-            } else {
-                sb.append(getVendor());
-            }
-            sb.append(" Family ");
-            sb.append(getFamily());
-            sb.append(" Model ");
-            sb.append(getModel());
-            sb.append(" Stepping ");
-            sb.append(getStepping());
-            this.cpuIdentifier = sb.toString();
-        }
-        return this.cpuIdentifier;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setIdentifier(String identifier) {
-        this.cpuIdentifier = identifier;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean isCpu64bit() {
-        return this.cpu64.booleanValue();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setCpu64(boolean value) {
-        this.cpu64 = Boolean.valueOf(value);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getStepping() {
-        return this.cpuStepping;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setStepping(String stepping) {
-        this.cpuStepping = stepping;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getModel() {
-        return this.cpuModel;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setModel(String model) {
-        this.cpuModel = model;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public String getFamily() {
-        return this.cpuFamily;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public void setFamily(String family) {
-        this.cpuFamily = family;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public synchronized double getSystemCpuLoadBetweenTicks() {
-        // Check if > ~ 0.95 seconds since last tick count.
-        long now = System.currentTimeMillis();
-        LOG.trace("Current time: {}  Last tick time: {}", now, tickTime);
-        if (now - tickTime > 950) {
-            // Enough time has elapsed.
-            updateSystemTicks();
-        }
-        // Calculate total
-        long total = 0;
-        for (int i = 0; i < curTicks.length; i++) {
-            total += (curTicks[i] - prevTicks[i]);
-        }
-        // Calculate idle from last field [3]
-        long idle = curTicks[3] - prevTicks[3];
-        LOG.trace("Total ticks: {}  Idle ticks: {}", total, idle);
-
-        return (total > 0 && idle >= 0) ? (double) (total - idle) / total : 0d;
     }
 
     /**
@@ -476,41 +265,6 @@ public class LinuxCentralProcessor implements CentralProcessor {
     };
 
     /**
-     * Updates system tick information from parsing /proc/stat. Stores in array
-     * with four elements representing clock ticks or milliseconds (platform
-     * dependent) spent in User (0), Nice (1), System (2), and Idle (3) states.
-     * By measuring the difference between ticks across a time interval, CPU
-     * load over that interval may be calculated.
-     */
-    private void updateSystemTicks() {
-        LOG.trace("Updating System Ticks");
-        // Copy to previous
-        System.arraycopy(curTicks, 0, prevTicks, 0, curTicks.length);
-        this.tickTime = System.currentTimeMillis();
-        long[] ticks = getSystemCpuLoadTicks();
-        System.arraycopy(ticks, 0, curTicks, 0, ticks.length);
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public double getSystemCpuLoad() {
-        if (sunMXBean) {
-            return ((com.sun.management.OperatingSystemMXBean) OS_MXBEAN).getSystemCpuLoad();
-        }
-        return getSystemCpuLoadBetweenTicks();
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public double getSystemLoadAverage() {
-        return getSystemLoadAverage(1)[0];
-    }
-
-    /**
      * {@inheritDoc}
      */
     @Override
@@ -530,34 +284,6 @@ public class LinuxCentralProcessor implements CentralProcessor {
             }
         }
         return average;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public double[] getProcessorCpuLoadBetweenTicks() {
-        // Check if > ~ 0.95 seconds since last tick count.
-        long now = System.currentTimeMillis();
-        LOG.trace("Current time: {}  Last tick time: {}", now, procTickTime);
-        if (now - procTickTime > 950) {
-            // Enough time has elapsed.
-            // Update latest
-            updateProcessorTicks();
-        }
-        double[] load = new double[logicalProcessorCount];
-        for (int cpu = 0; cpu < logicalProcessorCount; cpu++) {
-            long total = 0;
-            for (int i = 0; i < this.curProcTicks[cpu].length; i++) {
-                total += (this.curProcTicks[cpu][i] - this.prevProcTicks[cpu][i]);
-            }
-            // Calculate idle from last field [3]
-            long idle = this.curProcTicks[cpu][3] - this.prevProcTicks[cpu][3];
-            LOG.trace("CPU: {}  Total ticks: {}  Idle ticks: {}", cpu, total, idle);
-            // update
-            load[cpu] = (total > 0 && idle >= 0) ? (double) (total - idle) / total : 0d;
-        }
-        return load;
     }
 
     /**
@@ -600,27 +326,6 @@ public class LinuxCentralProcessor implements CentralProcessor {
             LOG.error("Problem with /proc/stat: {}", e.getMessage());
         }
         return ticks;
-    }
-
-    /**
-     * Updates per-processor tick information from parsing /proc/stat. Stores in
-     * 2D array; an array for each logical processor with four elements
-     * representing clock ticks or milliseconds (platform dependent) spent in
-     * User (0), Nice (1), System (2), and Idle (3) states. By measuring the
-     * difference between ticks across a time interval, CPU load over that
-     * interval may be calculated.
-     */
-    private void updateProcessorTicks() {
-        LOG.trace("Updating Processor Ticks");
-        // Copy to previous
-        for (int cpu = 0; cpu < logicalProcessorCount; cpu++) {
-            System.arraycopy(curProcTicks[cpu], 0, prevProcTicks[cpu], 0, curProcTicks[cpu].length);
-        }
-        this.procTickTime = System.currentTimeMillis();
-        long[][] ticks = getProcessorCpuLoadTicks();
-        for (int cpu = 0; cpu < logicalProcessorCount; cpu++) {
-            System.arraycopy(ticks[cpu], 0, curProcTicks[cpu], 0, ticks[cpu].length);
-        }
     }
 
     /**
@@ -684,22 +389,6 @@ public class LinuxCentralProcessor implements CentralProcessor {
      * {@inheritDoc}
      */
     @Override
-    public int getLogicalProcessorCount() {
-        return this.logicalProcessorCount;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public int getPhysicalProcessorCount() {
-        return this.physicalProcessorCount;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
     public int getProcessCount() {
         // Get all filenames in /proc directory with only digits (pids)
         File procdir = new File("/proc");
@@ -729,57 +418,5 @@ public class LinuxCentralProcessor implements CentralProcessor {
             LOG.error("Failed to get procs from sysinfo. {}", e.getMessage());
         }
         return 0;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public JsonObject toJSON() {
-        JsonArrayBuilder systemLoadAverageArrayBuilder = jsonFactory.createArrayBuilder();
-        for (double avg : getSystemLoadAverage(3)) {
-            systemLoadAverageArrayBuilder.add(avg);
-        }
-        JsonArrayBuilder systemCpuLoadTicksArrayBuilder = jsonFactory.createArrayBuilder();
-        for (long ticks : getSystemCpuLoadTicks()) {
-            systemCpuLoadTicksArrayBuilder.add(ticks);
-        }
-        JsonArrayBuilder processorCpuLoadBetweenTicksArrayBuilder = jsonFactory.createArrayBuilder();
-        for (double load : getProcessorCpuLoadBetweenTicks()) {
-            processorCpuLoadBetweenTicksArrayBuilder.add(load);
-        }
-        JsonArrayBuilder processorCpuLoadTicksArrayBuilder = jsonFactory.createArrayBuilder();
-        for (long[] procTicks : getProcessorCpuLoadTicks()) {
-            JsonArrayBuilder processorTicksArrayBuilder = jsonFactory.createArrayBuilder();
-            for (long ticks : procTicks) {
-                processorTicksArrayBuilder.add(ticks);
-            }
-            processorCpuLoadTicksArrayBuilder.add(processorTicksArrayBuilder.build());
-        }
-        JsonArrayBuilder systemIrqTicksArrayBuilder = jsonFactory.createArrayBuilder();
-        for (long ticks : getSystemIrqTicks()) {
-            systemIrqTicksArrayBuilder.add(ticks);
-        }
-        return NullAwareJsonObjectBuilder.wrap(jsonFactory.createObjectBuilder()).add("name", getName())
-                .add("physicalProcessorCount", getPhysicalProcessorCount())
-                .add("logicalProcessorCount", getLogicalProcessorCount())
-                .add("systemSerialNumber", getSystemSerialNumber()).add("vendor", getVendor())
-                .add("vendorFreq", getVendorFreq()).add("cpu64bit", isCpu64bit()).add("family", getFamily())
-                .add("model", getModel()).add("stepping", getStepping())
-                .add("systemCpuLoadBetweenTicks", getSystemCpuLoadBetweenTicks())
-                .add("systemCpuLoadTicks", systemCpuLoadTicksArrayBuilder.build())
-                .add("systemCpuLoad", getSystemCpuLoad()).add("systemLoadAverage", getSystemLoadAverage())
-                .add("systemLoadAverages", systemLoadAverageArrayBuilder.build())
-                .add("systemIOWaitTicks", getSystemIOWaitTicks())
-                .add("systemIrqTicks", systemIrqTicksArrayBuilder.build())
-                .add("processorCpuLoadBetweenTicks", processorCpuLoadBetweenTicksArrayBuilder.build())
-                .add("processorCpuLoadTicks", processorCpuLoadTicksArrayBuilder.build())
-                .add("systemUptime", getSystemUptime()).add("processes", getProcessCount())
-                .add("threads", getThreadCount()).build();
-    }
-
-    @Override
-    public String toString() {
-        return getName();
     }
 }
